Move SimpleRegion and make_src_region_file to lib
